Deduplicate common structs across instance data headers into shared common.hpp
