+++ projects/miopen/src/kernels/ck_builder/include/miopen/ck_builder/instance_data/xdl_v3.hpp
@@ -16,16 +16,7 @@ struct XdlV3Algorithm
     using GemmSpecial = ckb::GemmSpecialization;
     using PipeSched   = ckb::PipelineScheduler;
 
-    struct ThreadBlock
-    {
-        std::size_t block_size;
-        struct TileSize
-        {
-            std::size_t m;
-            std::size_t n;
-            std::size_t k;
-        } tile_size;
-    } thread_block;
+    ThreadBlock thread_block;
 
     struct GridwiseGemm
     {
@@ -40,52 +31,7 @@ struct XdlV3Algorithm
         } xdl_params;
     } gridwise_gemm;
 
-    struct TransferABC
-    {
-        struct TransferAB
-        {
-            struct BlockTransfer
-            {
-                std::size_t k0;
-                std::size_t m_n;
-                std::size_t k1;
-            } block_transfer;
-            struct LdsTransfer
-            {
-                std::size_t src_vector_dim;
-                std::size_t src_scalar_per_vector;
-                std::size_t lds_dst_scalar_per_vector;
-                bool is_direct_load;
-                bool lds_padding;
-            } lds_transfer;
-            struct BlockTransferAccessOrder
-            {
-                std::array<size_t, 3> order{0, 2, 1};
-            } thread_cluster_arrange_order;
-            struct SrcAccessOrder
-            {
-                std::array<size_t, 3> order{0, 2, 1};
-            } src_access_order;
-        };
-        TransferAB a;
-        TransferAB b;
-        struct TransferC
-        {
-            struct ThreadClusterDims
-            {
-                std::size_t m_block;
-                std::size_t m_wave_per_xdl;
-                std::size_t n_block;
-                std::size_t n_wave_per_xdl;
-            } thread_cluster_dims;
-            struct Epilogue
-            {
-                std::size_t m_xdl_per_wave_per_shuffle;
-                std::size_t n_per_wave_per_shuffle;
-                std::size_t scalar_per_vector;
-            } epilogue;
-        } c;
-    } transfer;
+    TransferABC transfer;
 
     ConvSpecial fwd_specialization;
     GemmSpecial gemm_specialization;
@@ -103,16 +49,7 @@ struct XdlV3Algorithm
 
     bool direct_load;
 
-    // Elementwise operations applied during convolution
-    // - input_op: Applied to input tensor (A) before GEMM
-    // - weight_op: Applied to weight tensor (B) before GEMM
-    // - output_op: Applied to output tensor (CDE) after GEMM (epilogue)
-    struct ElementwiseOps
-    {
-        ckb::ElementwiseOperation input_op;
-        ckb::ElementwiseOperation weight_op;
-        ckb::ElementwiseOperation output_op;
-    } elementwise_ops;
+    ElementwiseOps elementwise_ops;
 };
 
 static_assert(ckb::factory::FwdXdlV3Algorithm<XdlV3Algorithm>);
